refactor(web): replace CompassApi with BaseApi and remove unused compass.api files
* refactor: update API calls in various modules to use BaseApi instead of CompassApi
* chore: remove compass.api and related utility files
* feat: introduce base API structure for improved request handling and error management
* test: add unit tests for new base API functionality
