+++ packages/web/src/common/apis/auth.api.ts
@@ -4,11 +4,11 @@ import {
   type Result_Auth_Compass,
 } from "@core/types/auth.types";
 import { type GoogleAuthConfig } from "@web/auth/google/hooks/googe.auth.types";
-import { CompassApi } from "@web/common/apis/compass.api";
+import { BaseApi } from "@web/common/apis/base/base.api";
 
 const AuthApi = {
   async loginOrSignup(data: GoogleAuthConfig): Promise<Result_Auth_Compass> {
-    const response = await CompassApi.post<Result_Auth_Compass>(
+    const response = await BaseApi.post<Result_Auth_Compass>(
       `/signinup`,
       data,
       { headers: { rid: "thirdparty" } },
@@ -20,7 +20,7 @@ const AuthApi = {
   async connectGoogle(
     data: GoogleAuthCodeRequest,
   ): Promise<GoogleConnectResponse> {
-    const response = await CompassApi.post<GoogleConnectResponse>(
+    const response = await BaseApi.post<GoogleConnectResponse>(
       `/auth/google/connect`,
       data,
     );
