+++ packages/web/src/ducks/events/event.api.ts
@@ -4,15 +4,15 @@ import {
   type RecurringEventUpdateScope,
   type Schema_Event,
 } from "@core/types/event.types";
-import { type ApiResponse, CompassApi } from "@web/common/apis/compass.api";
+import { type ApiResponse, BaseApi } from "@web/common/apis/base/base.api";
 import { type Response_HttpPaginatedSuccess } from "@web/common/types/api.types";
 
 const EventApi = {
   create: (event: Schema_Event | Schema_Event[]) => {
-    return CompassApi.post<void>(`/event`, event);
+    return BaseApi.post<void>(`/event`, event);
   },
   delete: (_id: string, applyTo?: RecurringEventUpdateScope) => {
-    return CompassApi.delete<void>(`/event/${_id}?applyTo=${applyTo}`);
+    return BaseApi.delete<void>(`/event/${_id}?applyTo=${applyTo}`);
   },
   edit: (
     _id: string,
@@ -20,27 +20,27 @@ const EventApi = {
     params: { applyTo?: RecurringEventUpdateScope },
   ): Promise<ApiResponse<void>> => {
     if (params?.applyTo) {
-      return CompassApi.put<void>(
+      return BaseApi.put<void>(
         `/event/${_id}?applyTo=${params.applyTo}`,
         event,
       );
     }
 
-    return CompassApi.put<void>(`/event/${_id}`, event);
+    return BaseApi.put<void>(`/event/${_id}`, event);
   },
   get: (params: Params_Events) => {
     if (params.someday) {
-      return CompassApi.get<Response_HttpPaginatedSuccess<Schema_Event[]>>(
+      return BaseApi.get<Response_HttpPaginatedSuccess<Schema_Event[]>>(
         `/event?someday=true&start=${params.startDate}&end=${params.endDate}`,
       );
     } else {
-      return CompassApi.get<Response_HttpPaginatedSuccess<Schema_Event[]>>(
+      return BaseApi.get<Response_HttpPaginatedSuccess<Schema_Event[]>>(
         `/event?start=${params.startDate}&end=${params.endDate}`,
       );
     }
   },
   reorder: (order: Payload_Order[]) => {
-    return CompassApi.put(`/event/reorder`, order);
+    return BaseApi.put(`/event/reorder`, order);
   },
 };
